+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for drag and drop on the `ImageField` component.
+
 ## [10.0.0] - 2018-12-11
 ### Changed
 - The `Calendar` component `onChange` signature has changed to `onChange(year, month, day)`. This makes more sense when working with a calendar because it shouldn't be returning a date that is linked to a specific time zone.